Moved documentBuilder related things to separate namespace; Changed default configs
