+++ models/classes/search/index/IndexService.php
@@ -24,6 +24,7 @@ declare(strict_types=1);
 namespace oat\tao\model\search\index;
 
 use oat\oatbox\service\ConfigurableService;
+use oat\tao\model\search\index\DocumentBuilder\DocumentBuilderFactoryInterface;
 use oat\tao\model\search\Search;
 use oat\tao\model\menu\MenuService;
 use oat\generis\model\OntologyAwareTrait;
